timeout declared as unsigned eliminating number of warnings

+++ include/boost/test/impl/execution_monitor.ipp
@@ -230,7 +230,7 @@ report_error( execution_exception::error_code ec, boost::exception const* be, ch
     va_end( *args );
 
     throw execution_exception( ec, buf, execution_exception::location( extract<throw_file>( be ),
-                                                                       extract<throw_line>( be ),
+                                                                       (size_t)extract<throw_line>( be ),
                                                                        extract<throw_function>( be ) ) );
 }
 
@@ -676,7 +676,7 @@ signal_action::~signal_action()
 class signal_handler {
 public:
     // Constructor
-    explicit signal_handler( bool catch_system_errors, bool detect_fpe, int timeout, bool attach_dbg, char* alt_stack );
+    explicit signal_handler( bool catch_system_errors, bool detect_fpe, unsigned timeout, bool attach_dbg, char* alt_stack );
 
     // Destructor
     ~signal_handler();
@@ -699,7 +699,7 @@ public:
 private:
     // Data members
     signal_handler*         m_prev_handler;
-    int                     m_timeout;
+    unsigned                m_timeout;
 
     // Note: We intentionality do not catch SIGCHLD. Users have to deal with it themselves
     signal_action           m_ILL_action;
@@ -723,7 +723,7 @@ signal_handler* signal_handler::s_active_handler = signal_handler_ptr();
 
 //____________________________________________________________________________//
 
-signal_handler::signal_handler( bool catch_system_errors, bool detect_fpe, int timeout, bool attach_dbg, char* alt_stack )
+signal_handler::signal_handler( bool catch_system_errors, bool detect_fpe, unsigned timeout, bool attach_dbg, char* alt_stack )
 : m_prev_handler( s_active_handler )
 , m_timeout( timeout )
 , m_ILL_action ( SIGILL , catch_system_errors, attach_dbg, alt_stack )
@@ -781,8 +781,7 @@ signal_handler::~signal_handler()
 
     sigstk.ss_size  = MINSIGSTKSZ;
     sigstk.ss_flags = SS_DISABLE;
-    if( ::sigaltstack( &sigstk, 0 ) == -1 )
-    {
+    if( ::sigaltstack( &sigstk, 0 ) == -1 ) {
         int error_n = errno;
         std::cerr << "******** errors disabling the alternate stack:" << std::endl
                   << "\t#error:" << error_n << std::endl
@@ -886,13 +885,13 @@ public:
     {}
 
     void                report() const;
-    int                 operator()( unsigned int id, _EXCEPTION_POINTERS* exps );
+    int                 operator()( unsigned id, _EXCEPTION_POINTERS* exps );
 
 private:
     // Data members
     execution_monitor*  m_em;
 
-    unsigned int        m_se_id;
+    unsigned            m_se_id;
     void*               m_fault_address;
     bool                m_dir;
 };
@@ -901,7 +900,7 @@ private:
 
 #if BOOST_WORKAROUND( BOOST_MSVC, <= 1310)
 static void
-seh_catch_preventer( unsigned int /* id */, _EXCEPTION_POINTERS* /* exps */ )
+seh_catch_preventer( unsigned /* id */, _EXCEPTION_POINTERS* /* exps */ )
 {
     throw;
 }
@@ -910,9 +909,9 @@ seh_catch_preventer( unsigned int /* id */, _EXCEPTION_POINTERS* /* exps */ )
 //____________________________________________________________________________//
 
 int
-system_signal_exception::operator()( unsigned int id, _EXCEPTION_POINTERS* exps )
+system_signal_exception::operator()( unsigned id, _EXCEPTION_POINTERS* exps )
 {
-    const unsigned int MSFT_CPP_EXCEPT = 0xE06d7363; // EMSC
+    const unsigned MSFT_CPP_EXCEPT = 0xE06d7363; // EMSC
 
     // C++ exception - allow to go through
     if( id == MSFT_CPP_EXCEPT )
@@ -1087,7 +1086,7 @@ void BOOST_TEST_CALL_DECL
 invalid_param_handler( wchar_t const* /* expr */,
                        wchar_t const* /* func */,
                        wchar_t const* /* file */,
-                       unsigned int   /* line */,
+                       unsigned       /* line */,
                        uintptr_t      /* reserved */)
 {
     detail::report_error( execution_exception::user_error,
